refactor(shared): move interfaces out of constants files
Relocate the five interfaces declared in constants files into the
matching .interface.ts, per the LFXV2-2570 placement convention:

- CommitteeCategoryInfo -> interfaces/committee.interface.ts
- MeetingTypeConfig, MeetingTypeBadge -> interfaces/meeting.interface.ts
- TagTypeConfig -> interfaces/tag.interface.ts (new)
- TimezoneOption -> interfaces/timezones.interface.ts (new)

Constants files that type their config maps against a moved interface
re-import it type-only. Newly unused component-type imports are dropped.
The date-time util now imports TimezoneOption from the interfaces
barrel; the interfaces barrel gains the two new files. No behavior
change.

Refs: LFXV2-2573
