+++ packages/shared/src/constants/meeting.constants.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 import { ArtifactVisibility } from '../enums';
-import { TagSeverity } from '../interfaces';
+import type { MeetingTypeConfig } from '../interfaces/meeting.interface';
 import { lfxColors } from './colors.constants';
 
 /**
@@ -94,41 +94,6 @@ export const ARTIFACT_VISIBILITY_OPTIONS = [
   { label: 'Meeting Guests', value: ArtifactVisibility.MEETING_PARTICIPANTS },
   { label: 'Public', value: ArtifactVisibility.PUBLIC },
 ];
-
-/**
- * Meeting type color configuration
- * @description Defines colors, icons, and styling for meeting type badges and borders
- */
-export interface MeetingTypeConfig {
-  /** Display label for the meeting type */
-  label: string;
-  /** Background color class (e.g., bg-violet-100) */
-  bgColor: string;
-  /** Text color class - 600 shade (e.g., text-violet-600) */
-  textColor: string;
-  /** Text color class - 500 shade (e.g., text-violet-500) for alternate styling */
-  textColorAlt: string;
-  /** Border color class - 500 shade (e.g., border-violet-500) */
-  borderColor: string;
-  /** Border color class - 300 shade (e.g., border-violet-300) for lighter borders */
-  borderColorLight: string;
-  /** Font Awesome icon class */
-  icon: string;
-  /** CSS class for tag badge color override (e.g., tag-meeting-board) */
-  tagStyleClass: string;
-}
-
-/**
- * Meeting type badge interface
- * @description Structure for meeting type badge display
- */
-export interface MeetingTypeBadge {
-  label: string;
-  className: string;
-  severity?: TagSeverity;
-  styleClass?: string;
-  icon?: string;
-}
 
 /**
  * Meeting type color mappings
