+++ packages/shared/src/constants/tag.constants.ts
@@ -1,18 +1,8 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-import { TagProps, TagSeverity } from '../interfaces/components.interface';
-
-/**
- * Tag type configuration
- * @description Centralized tag styling configuration for common use cases
- */
-export interface TagTypeConfig {
-  severity: TagProps['severity'];
-  icon?: string;
-  rounded?: boolean;
-  styleClass?: string;
-}
+import { TagSeverity } from '../interfaces/components.interface';
+import type { TagTypeConfig } from '../interfaces/tag.interface';
 
 /**
  * Centralized tag type configurations
