+++ packages/shared/src/interfaces/committee.interface.ts
@@ -944,3 +944,14 @@ export interface CommitteeMemberPermissionInfo {
   /** True when `level` comes only from an inherited (project/foundation) grant, not a committee-scoped role. */
   inherited: boolean;
 }
+
+/**
+ * Committee category info for card-based selection
+ * @description Defines icons, descriptions, and colors for each committee category
+ */
+export interface CommitteeCategoryInfo {
+  icon: string;
+  description: string;
+  examples: string;
+  color: string;
+}
