+++ packages/shared/src/interfaces/meeting.interface.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 import { ArtifactVisibility, MeetingType, MeetingVisibility, RecurrenceType } from '../enums';
+import { TagSeverity } from './components.interface';
 
 // ============================================================================
 // V1 Legacy Summary Interfaces (still used by transformV1SummaryToV2)
@@ -1156,4 +1157,39 @@ export interface SummarySection {
   borderColor: string;
   /** Tailwind text color class for the icon */
   iconColor: string;
+}
+
+/**
+ * Meeting type color configuration
+ * @description Defines colors, icons, and styling for meeting type badges and borders
+ */
+export interface MeetingTypeConfig {
+  /** Display label for the meeting type */
+  label: string;
+  /** Background color class (e.g., bg-violet-100) */
+  bgColor: string;
+  /** Text color class - 600 shade (e.g., text-violet-600) */
+  textColor: string;
+  /** Text color class - 500 shade (e.g., text-violet-500) for alternate styling */
+  textColorAlt: string;
+  /** Border color class - 500 shade (e.g., border-violet-500) */
+  borderColor: string;
+  /** Border color class - 300 shade (e.g., border-violet-300) for lighter borders */
+  borderColorLight: string;
+  /** Font Awesome icon class */
+  icon: string;
+  /** CSS class for tag badge color override (e.g., tag-meeting-board) */
+  tagStyleClass: string;
+}
+
+/**
+ * Meeting type badge interface
+ * @description Structure for meeting type badge display
+ */
+export interface MeetingTypeBadge {
+  label: string;
+  className: string;
+  severity?: TagSeverity;
+  styleClass?: string;
+  icon?: string;
 }
